Update waitlist wording to "Get updates"
