+++ src/routes/waitlist/+page.server.ts
@@ -1,33 +1,6 @@
-import type { Actions } from './$types';
-import { env } from '$env/dynamic/private';
-import { error, redirect } from '@sveltejs/kit';
+import { redirect } from '@sveltejs/kit';
+import type { PageServerLoad } from './$types';
 
-export const actions = {
-    default: async ({ request }) => {
-        const data = await request.formData()
-
-        const email = data.get('email');
-        const team = data.get('team');
-
-        const body = {
-            "email": email,
-            "teamNumber": team,
-        }
-
-        if (env.WAITLIST_WEBHOOK == null) {
-            throw error(500, "We weren't able to sign you up.");
-        }
-
-        const response = await fetch(env.WAITLIST_WEBHOOK, {
-            method: "POST",
-            body: JSON.stringify(body),
-        });
-
-        if (await response.text() !== "Waitlist item recorded.") {
-            console.log(`Received an error when attempting to POST to the webhook: ${response.body}`);
-            throw error(500, "We weren't able to sign you up.");
-        }
-
-        throw redirect(303, '/waitlist/success');
-    },
-} satisfies Actions;
+export const load: PageServerLoad = () => {
+    throw redirect(301, '/get-updates');
+};
